backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
greenlet
ortools
pydantic-settings
python-dotenv
alembic
ortools
redis
celery
geopy
